requirements.txt
openai-whisper>=20231117
torch>=2.1.0
sounddevice>=0.4.6
numpy>=1.24.0
rich>=13.0.0
